NN: ongoing improvements on packaging.
git-svn-id: http://svn.zentyal.org/zentyal/trunk@21502 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ main/samba/debian/lucid/cdbs/ebox.mk
@@ -9,15 +9,14 @@ DEB_CONFIGURE_SCRIPT_ENV += IMAGESPATH="/usr/share/zentyal/www/images"
 DEB_CONFIGURE_SCRIPT_ENV += VARPATH="/var"
 DEB_CONFIGURE_SCRIPT_ENV += ETCPATH="/etc/zentyal"
 
-DEB_CONFIGURE_EXTRA_FLAGS := --disable-runtime-tests 
+DEB_CONFIGURE_EXTRA_FLAGS := --disable-runtime-tests
 DEB_MAKE_INVOKE = $(MAKE) $(DEB_MAKE_FLAGS) -C $(DEB_BUILDDIR)
 
 $(patsubst %,binary-install/%,$(DEB_PACKAGES)) :: binary-install/%:
 	for event in debian/*.upstart ; do \
-		if [ "x$$event" -ne "x" ] ; then \
-			install -d -m 755 debian/$(cdbs_curpkg)/etc/init; \
-			DESTFILE=$$(basename $$(echo $$event | sed 's/\.upstart/.conf/g')); \
-			install -m 644 "$$event" debian/$(cdbs_curpkg)/etc/init/$$DESTFILE; \
-		fi \
-	done;
-
+		[ -f $$event ] || continue; \
+		install -d -m 755 debian/$(cdbs_curpkg)/etc/init; \
+		DESTFILE=$$(basename $$(echo $$event | sed 's/\.upstart/.conf/g')); \
+		install -m 644 "$$event" debian/$(cdbs_curpkg)/etc/init/$$DESTFILE; \
+	done; \
+	chmod ugo+x debian/$(cdbs_curpkg)/etc/zentyal/hooks/*
